refactor(geometry): move arc-simplification out of the store clipping layer
Resolves issue #122 (design drift from PR #109).

Violation 1 — heavy geometry in the wrong layer: extract the arc/curve
reconstruction of Clipper output (known-circle reconstruction,
segment-preserving boolean reconstruction, and the Clipper-offset
simplification pipeline) out of src/store/helpers/clipping.ts into a new
pure-logic engine module, src/engine/toolpaths/arcReconstruction.ts.
clipping.ts is back to being a thin clipper-lib wrapper (profile↔path
conversion, boolean/offset execution, overlap predicates). The new module
imports only from engine/ and types/, so the store→engine dependency
direction is correct.

Violation 2 — sampling-constant duplication: export
DEFAULT_FLATTEN_CURVE_SAMPLES and DEFAULT_FLATTEN_ARC_STEP from
engine/toolpaths/geometry.ts, use them as the flattenProfile defaults, and
consume them in the annotation builder so the literals can no longer
diverge and silently break arc reconstruction.

Housekeeping: refresh src/store/INDEX.md and src/engine/INDEX.md, add the
arcReconstruction entry to toolpaths/INDEX.md, and add the missing
src/engine/simulation/INDEX.md (now 10 files, incl. gpuMesh/heightfieldShader).

Pure code motion — no behavior change. Build (tsc + structural tests + vite)
is green.
