+++ src/store/INDEX.md
@@ -13,7 +13,7 @@ Zustand store. The single source of truth for the current `.camj` project. **All
   - `pendingAddSlice.ts` — in-progress feature being drawn but not yet committed
   - `pendingCompletionSlice.ts` — partially-completed sketches awaiting closure
 - `helpers/` — pure helpers used by the store
-  - `clipping.ts` — clipper-lib wrappers (handles the integer scaling factor)
+  - `clipping.ts` — clipper-lib wrappers (handles the integer scaling factor): profile↔Clipper-path conversion, boolean/offset execution, and overlap predicates. Arc/curve reconstruction of Clipper output lives in `engine/toolpaths/arcReconstruction.ts`.
   - `derivedFeatures.ts` — computes derived features from the feature tree
   - `geometry.ts` — geometric utilities (bounds, transforms)
   - `ids.ts` — ID generation/uniqueness
